split stage and manager
