+++ ui/zenoedit/panel/ZLightsModel.cpp
@@ -7,6 +7,7 @@
 #include "zeno/utils/log.h"
 #include <zeno/types/UserData.h>
 #include <zenovis/ObjectsManager.h>
+#include <zenovis/StageManager.h>
 #include <zeno/types/PrimitiveObject.h>
 
 int ZLightsModel::rowCount(const QModelIndex &parent) const {
@@ -27,8 +28,10 @@ void ZLightsModel::updateByObjectsMan() {
     beginResetModel();
     light_names.clear();
     auto scene = Zenovis::GetInstance().getSession()->get_scene();
-    for (auto const &[key, ptr]: scene->objectsMan->lightObjects) {
-        if (ptr->userData().get2<int>("isL", 0)) {
+    // USD
+    //for (auto const &[key, ptr]: scene->objectsMan->lightObjects) {
+    for (auto const &[key, ptr]: scene->stageMan->zenoLightObjects) {
+        if (ptr->userData().get2<int>("isRealTimeObject", 0)) {
             light_names.push_back(key);
         }
     }
