+++ ui/zenoedit/viewport/viewporttransform.cpp
@@ -3,6 +3,7 @@
 #include <zeno/funcs/PrimitiveTools.h>
 #include <zeno/types/UserData.h>
 #include <zenovis/ObjectsManager.h>
+#include <zenovis/StageManager.h>
 #include <zenomodel/include/nodesmgr.h>
 
 #include <glm/gtx/transform.hpp>
@@ -37,7 +38,9 @@ FakeTransformer::FakeTransformer(const std::unordered_set<std::string>& names)
 
 void FakeTransformer::addObject(const std::string& name) {
     auto scene = Zenovis::GetInstance().getSession()->get_scene();
-    auto object = dynamic_cast<PrimitiveObject*>(scene->objectsMan->get(name).value());
+    // USD
+    //auto object = dynamic_cast<PrimitiveObject*>(scene->objectsMan->get(name).value());
+    auto object = dynamic_cast<PrimitiveObject*>(scene->stageMan->get(name).value());
     m_objects_center *= m_objects.size();
     auto& user_data = object->userData();
     zeno::vec3f bmin, bmax;
