+++ zenovis/include/zenovis/StageManager.h
@@ -1,22 +1,5 @@
 #pragma once
 
-// Include first for avoid error WinSock.h has already been included
-#include <boost/asio.hpp>
-#include <boost/bind.hpp>
-#include <boost/enable_shared_from_this.hpp>
-
-#include <boost/predef/os.h>
-#include <pxr/base/gf/camera.h>
-#include <pxr/base/js/json.h>
-#include <pxr/base/plug/plugin.h>
-#include <pxr/base/tf/fileUtils.h>
-#include <pxr/pxr.h>
-#include <pxr/usd/usdGeom/xform.h>
-#include <pxr/usd/usd/inherits.h>
-#include <pxr/usd/usd/stage.h>
-#include <pxr/usd/usdGeom/xformable.h>
-#include <pxr/usd/usdGeom/camera.h>
-
 #include <zenovis/Scene.h>
 #include <zeno/core/IObject.h>
 #include <zeno/utils/logger.h>
@@ -31,57 +14,42 @@
 #include <thread>
 #include <functional>
 
-PXR_NAMESPACE_USING_DIRECTIVE
+struct ZenoStage;
 
 namespace zenovis {
 
-struct ConfigurationInfo{
-    std::string cRepo = "http://test1:12345@192.168.3.11:8000/r/zeno_usd_test.git";
-    std::string cPath = "C:/Users/Public/zeno_usd_test";
-    std::string cRoot = "test.usda";
-    std::string cGit = "C:/Users/AMD/scoop/shims";
-    std::string cServer = "192.168.3.11";
-};
-
-struct PrimInfo{
-    std::string pPath;
-    std::shared_ptr<zeno::IObject> iObject;
-};
-
 struct StageManager : zeno::disable_copy {
+    /// ZenoObject (Editor) ---> UsdObject ---> ZenoObject (Convert)
+
+    // ZenoObject (Editor)
     zeno::MapStablizer<zeno::PolymorphicMap<
         std::map<std::string, std::shared_ptr<zeno::IObject>>>> zenoObjects;
 
-    UsdStageRefPtr cStagePtr;
-    UsdStageRefPtr sStagePtr;
-    ConfigurationInfo confInfo;
+    // ZenoObject (Convert)
+    zeno::MapStablizer<zeno::PolymorphicMap<
+        std::map<std::string, std::shared_ptr<zeno::IObject>>>> convertObjects;
 
-    std::string pathEnv;
+    // ZenoObject - Light
+    std::map<std::string, std::shared_ptr<zeno::IObject>> zenoLightObjects;
+
+    std::shared_ptr<ZenoStage> zenoStage;
+
+    int increase_count = 0;
 
     StageManager();
     ~StageManager();
 
     template <class T = void>
     auto pairs() const {
+        // XXX
         return zenoObjects.pairs<T>();
     }
-
     template <class T = void>
     auto pairsShared() const {
         return zenoObjects.pairsShared<T>();
     }
 
     bool load_objects(std::map<std::string, std::shared_ptr<zeno::IObject>> const &objs);
-    void update();
-
-    int _UsdGeomMesh(const PrimInfo& primInfo);
-
-    void _CreateUSDHierarchy(const SdfPath &path)
-    {
-        if (path == SdfPath::AbsoluteRootPath())
-            return;
-        _CreateUSDHierarchy(path.GetParentPath());
-        UsdGeomXform::Define(cStagePtr, path);
-    }
+    std::optional<zeno::IObject*> get(std::string nid);
 };
 }
